Relocate the Type file and remove unused code
